+++ Tests/Unit/BackEnd/AbstractEventMailFormTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use OliverKlee\Seminars\Tests\Unit\BackeEnd\Support\Traits\BackEndTestsTrait;
+use OliverKlee\Seminars\Tests\Unit\Support\Traits\BackEndTestsTrait;
 use TYPO3\CMS\Backend\Utility\BackendUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
@@ -53,10 +53,6 @@ class Tx_Seminars_Tests_Unit_BackEnd_AbstractEventMailFormTest extends Tx_Phpuni
     {
         $this->unifyTestingEnvironment();
 
-        $configuration = new Tx_Oelib_Configuration();
-        Tx_Oelib_ConfigurationRegistry::getInstance()->set('plugin.tx_seminars', $configuration);
-
-        Tx_Oelib_HeaderProxyFactory::getInstance()->enableTestMode();
         /** @var Tx_Oelib_MailerFactory $mailerFactory */
         $mailerFactory = GeneralUtility::makeInstance(Tx_Oelib_MailerFactory::class);
         $mailerFactory->enableTestMode();
@@ -734,7 +730,7 @@ class Tx_Seminars_Tests_Unit_BackEnd_AbstractEventMailFormTest extends Tx_Phpuni
                 false,
                 true
             ),
-            Tx_Oelib_HeaderProxyFactory::getInstance()->getHeaderProxy()->getLastAddedHeader()
+            $this->headerProxy->getLastAddedHeader()
         );
     }